Fixing the issue-2
From the elemental level, write a new function crtstif_enriched_elemental to guide/direct the building of Jacobian Matrix given that the shape functions and derivatives at the gaussian points are already enriched by every enrichitem within the element.

This commit also include one or two places which should be included in the fix of issue-1.

The change in elem.calstress_enriched now use the closest 3 but not 2 gaussian points to estimate the contribution to node stress.