#21 update nodespool and phipool within realgrow
Issue: within the update_enrich_1, the call of intersection method to assign localpnts to newelem, is using the empty varargin. This will trigger the use Mygeo.Nodespool and Phipool. Somehow, these two info was unnecessary in before this project and was not updated, causing problem.

Fix: update mygeo inside EnrCrackTip.realgrow.