One small change to fix the pressure blocking by the intersection element
Refer to isssue #19

The two cracks in the intersection element seems to be disconnected when it comes to fluid penetration although they geometrically connected.
I guess it is because the domain permeability is too small and the rigde enrichment just creates two ridged pore pressure profile in the intersection element. The fracture fluid flow only has a big perm along the tangential direction.

Therefore, I think setting the intersection element a big domain kmat may force the pressure exchange between the two cracks.

Now when there are more than 1 cracks in one element, kmat is set to 1000 which is about the same magnitude of crack perm.