Change crackbody.perforated to InitialMode issue #24
# The explanation on InitialMode
1: perforated, completely traction free;
2: "smeared crack" or cemented crack, seemingly continuum
3:existing fracture, start with compressive mode
4: newly propagated segment, start with tensile mode.

Major changes are in GaussPnt_Cohesive.initiate. By setting inicrackdisp a very small negative value, the tractionlaw.gvietangent will give the compressive tangent mode.

# Observations:
I tried the compressive mode, but upon initial discharge, the crack already starts opening, which is not good. When I tried tensile mode, (tangent =0 when kfirst=0)it is not convergent again. Not good.

If the insitu stress is assigned as nonactive load, it is not very practical to introduce this compressive mode or tensile mode.

# Next step:
Going directly to the modeling of 'smeared crack'.