The divergence problem is fixed temporarily
It turns out the divergence problem in 898db3386e9ae8154600b580e4d1f9d8b1897fe5 is gone when the tini and tkrg for the normal crack is not much larger than those for the opened smeared crack. I have not figured out the exact reason behind it. But for now, the code can run beautifully before the opened crack go back to compressive mode. That would be the final part of Issue #38 .