The follow up of divergence problem of commit 6b9b561008b1f30a557181f983f1c5e86d831309
# Suspect cause
The stress calculation for the "smeared" element right ahead of the propagating crack may be unrealistically high. So it is become quite difficult for the system to converge if we assign the traction to tractionO using that stress prior to smeared crack open.  While the real reason is still unclear, the problem seems to be fixed by one simplification:

# Simplified solution:

Make the transited smeared crack traction-less. The solution is done by
1. Make tini and tkrg for these smeared crack 0 initially.
2. Make the traction)=[0;0] in GaussPnt_Cohesive.transit.
3. Additionally, it is noticed that when the tip is approaching the unsmeared crack, the maxps using the "center" mode becomes unreliable. So we change the mode to tip. Since it is not very important to achieve crack propagation early or late, the Maxpscheck.Tol2 (used to check unstable) is muted. so that the crack growth will be always allowed when `Maxpscheck.Dimensionless>1`.